% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_control}
\alias{add_control}
\title{Add a control to the map}
\usage{
add_control(map, name = "NavigationControl", position = "top-left",
  ...)
}
\arguments{
\item{map}{map widget created by \link{rvl}}

\item{name}{name of the control}

\item{position}{position on the map}

\item{...}{options passed to the control}
}
\description{
Add a control to the map
}
