% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carto-dataset-layer.R
\name{add_dataset_layer}
\alias{add_dataset_layer}
\title{Add a (CARTO) dataset layer to the map}
\usage{
add_dataset_layer(map, dataset_name, viz_def = list(),
  id = "dataset-layer", ...)
}
\arguments{
\item{map}{map widget created by \link{cartovl}}

\item{dataset_name}{name of the dataset which will be add to the map}

\item{viz_def}{visualization definition, character vector; if set to \code{NULL}
a default style will be applied}

\item{id}{unique layer id}

\item{...}{more props}
}
\description{
Add a (CARTO) dataset layer to the map
}
