% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{add_external_geojson_layer}
\alias{add_external_geojson_layer}
\title{Add an external layer to the map}
\usage{
add_external_geojson_layer(map, source, viz_def = list(),
  id = "external-layer", ...)
}
\arguments{
\item{map}{map widget created by \link{cartovl}}

\item{source}{url of the geojson data source}

\item{viz_def}{visualization definition, character vector; if set to \code{NULL}
a default style will be applied}

\item{id}{unique layer id}

\item{...}{more props}
}
\description{
Add an external layer to the map
}
