% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapboxgl-stuff.R
\name{add_mapbox_layer}
\alias{add_mapbox_layer}
\title{Add a layer using the mapbox style specification to the map}
\usage{
add_mapbox_layer(map, source_id, style, id = "mapbox-layer")
}
\arguments{
\item{map}{map widget created by \link{rvl}}

\item{source_id}{The id of the source to get data from.}

\item{style}{style conforming to the mapbox style specification}

\item{id}{unique id of the style, will be added to \code{style}}
}
\description{
Add a layer using the mapbox style specification to the map
}
