% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_navigation_control}
\alias{add_navigation_control}
\title{Add navigation control to the map}
\usage{
add_navigation_control(map, position = "top-left", ...)
}
\arguments{
\item{map}{map widget created by \link{cartovl}}

\item{position}{position on the map}

\item{...}{more options passed to the navigation control}
}
\description{
Add navigation control to the map
}
