% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-tiles.R
\name{add_raster_tiles}
\alias{add_raster_tiles}
\title{Add raster tiles to the map}
\usage{
add_raster_tiles(map, tiles = NULL, attribution = NULL)
}
\arguments{
\item{tiles}{tile sources}

\item{attribution}{attribution}
}
\description{
Add raster tiles to the map
}
