% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapboxgl-stuff.R
\name{add_source}
\alias{add_source}
\title{Add a source to the map}
\usage{
add_source(map, data, id = "source", ...)
}
\arguments{
\item{map}{map widget created by \link{Rvl}}

\item{data}{source object}

\item{id}{The id of the source to add.}

\item{...}{more props}
}
\description{
Add a source to the map
}
