% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartovl.R
\name{cartovl}
\alias{cartovl}
\title{Create a mapbox-gl / carto-vl (map) widget}
\usage{
cartovl(
  properties = map_properties(),
  width = NULL,
  height = NULL,
  minWidth = "480px",
  minHeight = "360px",
  element_id = NULL
)
}
\arguments{
\item{properties}{properties passed to \code{mapboxgl.Map}}

\item{width}{styled width of the widget}

\item{height}{styled height of the widget}

\item{minWidth}{minimum width}

\item{minHeight}{minimum height}

\item{element_id}{explicit element id (usually not needed)}
}
\description{
Create a mapbox-gl / carto-vl (map) widget
}
