% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapboxgl-stuff.R
\name{mapbox_text_style}
\alias{mapbox_text_style}
\title{Make a mapbox text style}
\usage{
mapbox_text_style(text_field, text_color = "white", text_size = 10)
}
\arguments{
\item{text_field}{Value to use for a text label.}

\item{text_color}{The color with which the text will be drawn.}

\item{text_size}{The size with which the text will be drawn. Units are in pixels.}
}
\description{
Make a mapbox text style
}
