% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartovl.R
\name{rvlOutput}
\alias{rvlOutput}
\title{Shiny bindings for carto-vl}
\usage{
rvlOutput(
  outputId,
  width = "100\%",
  height = "100\%",
  minWidth = "480px",
  minHeight = "360px"
)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width}{width of div in browser; must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{height}{height of div in browser; must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{minWidth}{minimum width}

\item{minHeight}{minimum height}
}
\description{
Output and render functions for using cartovl within Shiny
applications and interactive Rmd documents.
}
