% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carto-dataset-layer.R
\name{set_default_auth}
\alias{set_default_auth}
\title{Set default authentication parameters: user and API key}
\usage{
set_default_auth(map, user_name = "cartovl", api_key = "default_public")
}
\arguments{
\item{map}{map widget created by \link{Rvl}}

\item{user_name}{name of the user}

\item{api_key}{API key used to authenticate against CARTO.}
}
\description{
Set default authentication parameters: user and API key
}
