% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-properties.R
\name{set_view}
\alias{set_view}
\title{Set the view of the map}
\usage{
set_view(map, longitude, latitude, zoom = 2)
}
\arguments{
\item{map}{map widget created by \link{Rvl}}

\item{longitude}{longitude}

\item{latitude}{latitude}

\item{zoom}{zoom}
}
\description{
Set the view of the map
}
