% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{add_box}
\alias{add_box}
\title{Add a box to the map}
\usage{
add_box(map, html)
}
\arguments{
\item{map}{map widget created by \link{cartovl}}

\item{html}{The text which will be displayed on the map}
}
\description{
Add a box to the map
}
