% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{add_box}
\alias{add_box}
\title{Add a (text) box to the map}
\usage{
add_box(map, content, position = "top-right", style = "")
}
\arguments{
\item{map}{map widget created by \link{cartovl}}

\item{content}{The text which will be displayed on the map.}

\item{position}{The position on the map.}

\item{style}{CSS options in the form of \emph{cssText} applied to the element.}
}
\description{
Add a (text) box to the map
}
