% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{add_carto_basemap}
\alias{add_carto_basemap}
\title{Add a base map from CARTO to the map widget}
\usage{
add_carto_basemap(map, style_name = "darkmatter")
}
\arguments{
\item{map}{map widget created by \link{cartovl}}

\item{style_name}{map style, one of \code{darkmatter}, \code{positron} or \code{voyager}}
}
\description{
Add a base map from CARTO to the map widget
}
