% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{add_layer}
\alias{add_layer}
\title{Add layer to map widget}
\usage{
add_layer(map, data, id = "layer")
}
\arguments{
\item{data}{data (geojson)}

\item{id}{unique layer id}
}
\description{
Add layer to map widget
}
