% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-tiles.R
\name{add_raster_tiles}
\alias{add_raster_tiles}
\title{Add raster tiles to the map}
\usage{
add_raster_tiles(map, tiles = NULL, attribution = NULL)
}
\arguments{
\item{map}{map widget created by \link{cartovl}}

\item{tiles}{A vector of one or more tile source URLs,
e. g. \code{https://stamen-tiles.a.ssl.fastly.net/watercolor/{z}/{x}/{y}.jpg}}

\item{attribution}{The attribution which will be displayed on the map.}
}
\description{
Add raster tiles to the map
}
