% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartovl.R
\name{cartovl}
\alias{cartovl}
\title{Create a mapbox-gl / carto-vl (map) widget}
\usage{
cartovl(properties = list(), width = NULL, height = NULL,
  element_id = NULL)
}
\arguments{
\item{properties}{properties passed to \code{mapboxgl.Map}}

\item{width}{width of the widget}

\item{height}{height of the widget}

\item{element_id}{explicit element id (usually not needed)}
}
\description{
Create a mapbox-gl / carto-vl (map) widget
}
