% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-properties.R
\name{map_properties}
\alias{map_properties}
\title{Map properties}
\usage{

  map_properties(style = "https://basemaps.cartocdn.com/gl/dark-matter-gl-style/style.json",
  center = c(0, 0), zoom = 2)
}
\arguments{
\item{style}{basemap style}

\item{center}{longitude, latitude}

\item{zoom}{zoom}
}
\description{
Map properties
}
