% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-properties.R
\name{map_properties}
\alias{map_properties}
\title{Map properties / options}
\usage{
map_properties(style = NULL, min_zoom = NULL, max_zoom = NULL,
  center = c(0, 0), zoom = 2, ...)
}
\arguments{
\item{style}{The mapbox style of the map.}

\item{min_zoom}{The minimum zoom level of the map (0 - 24).}

\item{max_zoom}{The maximum zoom level of the map (0 - 24).}

\item{center}{The inital geographical centerpoint of the map (longitude, latitude coordinate order).}

\item{zoom}{The initial zoom level of the map.}

\item{...}{more properties, see \href{https://docs.mapbox.com/mapbox-gl-js/api/}{mabox-gl API reference}}
}
\description{
Properties like \code{center} or \code{zoom} are usually set via \link{set_view}.
The \code{style} can be set by functions like \link{add_carto_basemap} and others.
}
