% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-properties.R
\name{set_view}
\alias{set_view}
\title{Set the view of the map}
\usage{
set_view(map, longitude, latitude, zoom = 2, pitch = 0, bearing = 0)
}
\arguments{
\item{map}{map widget created by \link{cartovl}}

\item{longitude}{The longitude of the geographical centerpoint of the map.}

\item{latitude}{The latitude of the  geographical centerpoint of the map.}

\item{zoom}{The zoom level of the map.}

\item{pitch}{The pitch (tilt) of the map, measured in degrees away from the plane of the screen (0 - 60).}

\item{bearing}{The bearing (rotation) of the map, measured in degrees counter-clockwise from north.}
}
\description{
Set the view of the map
}
